Rearranged the folder structure
